requirement.txt
transformers
torch
sentence-transformers
scikit-learn
numpy
pandas
chromadb
langchain-schema
langchain-vectorstores
langchain-embeddings